bfpg-aeson: improve appearance of inline code
Display inline <code> in hot pink and use the CSS white-space: pre;
directive to ensure that we can still line things up.  <code>
elements that are descendents of <li> also get a subtle border.

Also adjust the whitespace in and around a handful of <code>
elements so that we have only what we want inside the border.